requirements.txt
# ============================================================================
# LLM Provider Dependencies
# ============================================================================
google-generativeai>=0.8.0
groq>=0.11.0
tenacity==8.2.3

# ============================================================================
# Document Processing
# ============================================================================
PyPDF2==3.0.1
python-docx==1.1.0

# ============================================================================
# Data Processing
# ============================================================================
pandas==2.1.4
pydantic==2.5.3

# ============================================================================
# Utilities
# ============================================================================
python-dotenv==1.0.0
pyyaml==6.0.1

# ============================================================================
# ML/NLP (for deduplication)
# ============================================================================
sentence-transformers==2.3.1
scikit-learn==1.4.0

# ============================================================================
# Testing
# ============================================================================
pytest==7.4.3

